Get Rust Virtio GPU driver working with Virtio PCI framework
