+++ src/dev/Makefile
@@ -14,7 +14,7 @@ obj-$(NAUT_CONFIG_HPET) += hpet.o
 obj-$(NAUT_CONFIG_VIRTIO_PCI) += virtio_pci.o
 obj-$(NAUT_CONFIG_VIRTIO_NET) += virtio_net.o
 obj-$(NAUT_CONFIG_VIRTIO_BLK) += virtio_blk.o
-obj-$(NAUT_CONFIG_VIRTIO_GPU) += virtio_gpu.o
+obj-$(NAUT_CONFIG_C_VIRTIO_GPU) += virtio_gpu.o
 
 obj-$(NAUT_CONFIG_E1000_PCI) += e1000_pci.o
 obj-$(NAUT_CONFIG_E1000E_PCI) += e1000e_pci.o
